+++ processor/src/test/fixtures/bad_input/com/example/inject/Jsr330InjectFieldDaggerModel.java
@@ -5,12 +5,10 @@ import arez.annotations.ArezComponent;
 import javax.inject.Inject;
 
 @ArezComponent
-public abstract class MethodInjectionModel
+public abstract class Jsr330InjectFieldDaggerModel
 {
   @Inject
-  void setService( Runnable myService )
-  {
-  }
+  Runnable _myService;
 
   @Action
   void myAction()